requirements.txt
python-telegram-bot==20.6
google-generativeai==0.3.1
requests==2.31.0